Implemented whole CRUDS fucntions in inndex  page
