+++ src/components/app/ModelSchema.ts
@@ -1,0 +1,16 @@
+const BandSchema = [
+  {
+    title: 'Name of Band',
+    name: 'name',
+    description: 'Enter name of band',
+    required: true,
+  },
+  {
+    title: 'Description of Band',
+    name: 'description',
+    description: 'Enter description of band',
+    required: false,
+  },
+];
+
+export { BandSchema };
